refactor: Split Bank type to Bank & BankManager
