+++ BankManagerConsoleApp/Model/Bank.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Bank {
-    private enum BankStatus: Int {
+    enum BankStatus: Int {
         case open = 1
         case close = 2
     }
@@ -25,11 +25,11 @@ struct Bank {
         }
     }
     
-    private func receiveCustomer() {
+    func receiveCustomer() {
         receiveCustomer(range: defaultFrom...defaultTo)
     }
     
-    private func doTask() {
+    func doTask() {
         var customer: Customer?
         while !queue.isEmpty {
             customer = queue.dequeue()
@@ -52,33 +52,5 @@ struct Bank {
         }
         let finishTime = Double(customer.id) * 0.7
         print("업무가 마감되었습니다. 오늘 업무를 처리한 고객은 총 \(customer.id)명이며, 총 업무시간은 \(String(format: "%.2f", finishTime)) 입니다.")
-    }
-}
-
-extension Bank: Process {
-    
-    private func choiceMenuWithGuide() -> BankStatus? {
-        print("1 : 은행개점")
-        print("2 : 종료")
-        print("입력 : ", terminator: "")
-        guard let userInput = readLine(), let filterdInput = Int(userInput) else {
-            return nil
-        }
-        return BankStatus(rawValue: filterdInput)
-    }
-    
-    func runProcess() {
-        while true {
-            let bankStatus = choiceMenuWithGuide()
-            switch bankStatus {
-            case .open:
-                receiveCustomer()
-                doTask()
-            case .close:
-                return
-            default:
-                print("잘못 입력했습니다. 다시 입력해주세요.")
-            }
-        }
     }
 }
