--- BankManagerConsoleApp/Model/Process.swift
@@ -1,12 +1,0 @@
-//
-//  Process.swift
-//  BankManagerConsoleApp
-//
-//  Created by kjs on 2021/07/30.
-//
-
-import Foundation
-
-protocol Process {
-    func runProcess()
-}
